Accumulated changes:
- Import statements now *must* be separated by semicolons. This makes
them consistent with args.

- StandardTemplateManager now becomes beanlike with a default
constructor and setters for work dir and source dir (which as it turns
out, have not-so-unreasonable defaults). Since we're in rapid
development mode with few users, the old constructors were removed
instead of deprecated.

- The TemplateGenerator gets the "smarts" (it's still horrible
cutnpaste code) to generate template Impls as well as interfaces.
Still to do: get the ant task up to speed on this.

- Revisit the tests, bringing them up to date with changes in the
StandardTemplateManager as well as refactoring them into fairly
reasonable shape.


git-svn-id: svn://svn.code.sf.net/p/jamon/code/trunk/jamon@186 6afd1a24-8710-0410-9bc4-c69fca332c4c