A stab at converting exposed checked exceptions to runtime exceptions, except for IOExceptions related to writing to the underyling Writer.
In the process, the Compiler tool classes don't throw an exception on
a failed compilation, but rather return a String result, null if
sucessful. JamonException becomes abstract.

What's incomplete here is that the reflection-based TemplateInspector
(and the clients InvokerTask and InvokerTool) ought to be able to
distinguish a "template not found" condition, and be more specific
about errors in general.


git-svn-id: svn://svn.code.sf.net/p/jamon/code/trunk/jamon@1406 6afd1a24-8710-0410-9bc4-c69fca332c4c