Checkpoint on the way to allowing the eclipse plugin to use the jamon jar in the project, rather than it's own copy:
*) ParserError, ParserErrors and Location become interfaces, implemented by a corresponding Impl class.
*) The jamon-specific builder code in the eclipse plugin is moved to JamonBuilderImpl (interfaced by JamonBuilder).

The idea here is to reflectively create a JamonBuilderImpl class from the project's class path, and then work with that, via interfaces (which can "cross" classloader boundaries).

All that done and said, this may be a dead end.  To fully integrate jamon into eclipse, we need to have eclipse doing parsing, and making use of the parse tree.  To do this using the projects jar would require (at least) an interface for every node, and even then, would be tied to a version of the syntax.


git-svn-id: svn://svn.code.sf.net/p/jamon/code/trunk/jamon@1767 6afd1a24-8710-0410-9bc4-c69fca332c4c